requirements.txt
click==8.1.3
ffmpy==0.3.0
python-dotenv==0.20.0
requests==2.28.1
spotipy==2.20.0
SQLAlchemy==1.4.40
tldextract==3.3.1
validators==0.20.0
yt_dlp==2022.8.19
